// web/expo-money/src/types/customerFilterDTO.tsx
export interface CustomerFilterDTO {
    id: string
    firsName: string
    lastName: string
    contact: string
    photo: string
    urlPhoto: string
    amountFinancialLoans: number
    amountFinancialLoansOpen: number
    amountFinancialLoansPending: number
    amountFinancialLoansExecutedPledge: number
}